backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
google-generativeai==0.3.1
celery==5.3.4
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
geopy==2.4.1
pandas==2.1.4

